lots more debug output; fixed(?) error when loading instrument
When we load an instrument OnIntrument(ev) is called, which loads the
corresponding xml node from config.xml and passes it to the Instrument
contructor. The constructor now reads the path from the config node and
loads a hierarchy of xml files representing the instrument (have a look
at the data description pdf from the original authors)
After loading the instrument, OnInstrument will try to instanciate a
PatchEditor with a dummy patch. This will fail if the patchType does not
match any patch described in any of the xml files linked in the data
field of the instrument node. This is where None was changed to 'single'
to see if we get any further.
NOTE: the tag names (e.g. patch) are not unambiguous, which can be quite
confusing (sometimes it's the description data, sometimes it's a link
and we have to trust the parser to assume the right meaning in the right
instance.
Also patch can refer to different types of data:
  - midi messages used to communicate with the device (PC, dump req, ..)
  - patch data descriptions (assign meaning to the bytes in the dump)
PatchEditor will now try to build a PatchEditTree from the information
in instrument, patch and the decoders defined for that instrument.
This is split in two phases:
  - LoadPatch (build the tree)
  - tree.expandTree (display the tree)
LoadPatch will try to load the matching Parameter set for the type of
editor requested from the xml description and assign widgets to them
according to the interface mode/type
ExpandTree will call Expand(node) on all nodes in self.allNodes, this is
a wxWidgets method inherited from wx.TreeCtrl

This is where our journey ends for this commit.