refactor and fix of tag.py
How do we find an error for which we only get a confusing error message?

AttributeError: HandleTag

This is all we got.

$ ag HandleTag

shows us there is a HandleTag in tag.py, which is never called from our
codebase. But it belongs to a class which inherits from
wx.lib.wxpTag.wxpTagHandler and is registered as a HTML TagParser.
We might suspect it could be a name clash because of all the
from ... import * lines and refactor everything to import ... (which is
a good thing anyway in this case) but get nowhere.

Our beloved friend binary search to the rescue!
We comment out either

return self.HandleWidgetTag(tag)

or

return self.HandleParameterTag(tag)

and notice the problem lies in HandleWidgetTag but the error message
gets lost in the logging/error printing mechanics of the wx library.
Without any further info where to look we insert print('mark n')
statements at strategic places (more than one, so this is not strictly
binary anymore but who cares ;) ) and see how far we get (mark 1).
There we do the same thing again to further reduce the region of code
that causes the error.
At first it seemd clever to encapsulate the small region of code in a
try-except-AttributeError block and print the error as that is what the
error message tells us is happening. But unfortunately no such error is
triggered.
As it turns out the library not only takes away the error message but
also silently changes the error type from TypeError (which is
explicitely raised in our code) to AttributeError to infuriate the
debugger.
Upon tracing the source of the error we notice that the author wants to
check if a dynamically loaded object (by name from xml) is a classobject
but the test fails even though what is imported/loaded seems to be a
class. A quick internet search later we learn that we can check if an
object is a class object by passing it to inspect.isclass. This works
just fine and returns True for all the objects. Maybe the other way
never worked (which I doubt) or the way to do this check in python2 has
changed in the last 10-15 years since the codebase was last touched.
We again get one step closer to reviving ancient code and I'm quite
surprised how well things still work - especially in the wxWidgets
world. But nonetheless the next error message is waiting for us:

'HtmlWinParser' object has no attribute 'GetWindow'

With this I wish a pleasant night! :)